Add support for setting permissions on dotfiles
SSH in particular requires that ~/.ssh/config is not group or world
read/writable (i.e. it should be 0600). Git does not record file
permissions.

This commit adds support for dotfiles_permissions.config in the dotfiles
repo. It specifies permissions for dotfiles. After a pull, the permissions
specified in this file will be applied.